Split packet generation code into its own module
