+++ aioesphomeapi/_frame_helper/plain_text.pxd
@@ -2,12 +2,10 @@ import cython
 
 from ..connection cimport APIConnection
 from .base cimport APIFrameHelper
+from .packets cimport make_plain_text_packets
 
-
-cdef object varuint_to_bytes
 cdef bint TYPE_CHECKING
 
-cpdef _varuint_to_bytes(cython.int value)
 
 cdef class APIPlaintextFrameHelper(APIFrameHelper):
 
@@ -24,10 +22,4 @@ cdef class APIPlaintextFrameHelper(APIFrameHelper):
     )
     cdef int _read_varuint(self)
 
-    @cython.locals(
-        type_="unsigned int",
-        data=bytes,
-        packet=tuple,
-        type_=object
-    )
     cpdef void write_packets(self, list packets, bint debug_enabled) except *
