+++ aioesphomeapi/_frame_helper/plain_text.py
@@ -1,34 +1,13 @@
 from __future__ import annotations
 
 import asyncio
-from functools import lru_cache
 from typing import TYPE_CHECKING
 
 from ..core import ProtocolAPIError, RequiresEncryptionAPIError
 from .base import APIFrameHelper
+from .packets import make_plain_text_packets
 
 _int = int
-
-
-def _varuint_to_bytes(value: _int) -> bytes:
-    """Convert a varuint to bytes."""
-    if value <= 0x7F:
-        return bytes((value,))
-
-    result = bytearray()
-    while value:
-        temp = value & 0x7F
-        value >>= 7
-        if value:
-            result.append(temp | 0x80)
-        else:
-            result.append(temp)
-
-    return bytes(result)
-
-
-_cached_varuint_to_bytes = lru_cache(maxsize=1024)(_varuint_to_bytes)
-varuint_to_bytes = _cached_varuint_to_bytes
 
 
 class APIPlaintextFrameHelper(APIFrameHelper):
@@ -48,17 +27,7 @@ class APIPlaintextFrameHelper(APIFrameHelper):
 
         The entire packet must be written in a single call.
         """
-        out: list[bytes] = []
-        for packet in packets:
-            type_: int = packet[0]
-            data: bytes = packet[1]
-            out.append(b"\0")
-            out.append(varuint_to_bytes(len(data)))
-            out.append(varuint_to_bytes(type_))
-            if data:
-                out.append(data)
-
-        self._write_bytes(out, debug_enabled)
+        self._write_bytes(make_plain_text_packets(packets), debug_enabled)
 
     def _read_varuint(self) -> _int:
         """Read a varuint from the buffer or -1 if the buffer runs out of bytes."""
